chore(alerts|kube-monitoring): managed prometheus-operator and CRDs
Installation of the CustomResourceDefinitions and the prometheus operator is a unique task for a cluster. We have two plugins that depend on it and therefore need to be able to bring the basis for their installation themselves. Therefore we decided to use the functionality of the kube-prometheus-stack to manage CRDs and turn it on by default. This makes chart testing considerably easier.

asdsa