feat(audit-logs):  Adds plugin of OpenTelemetry Audit Logs
This adds a plugin for OpenTelemetry ingestion of audit relevant signals. This will be a separate plugin from the 'OpenTelemetry' plugin, because it will most likely have different requirements.

This is an initial pull to create the foundation of the plugin.

---------

Signed off by: Simon Olander (simon.olander@sap.com)